#942 Creation assessments app + move score encoding from dashboard to new app
