+++ assessments/templates/score_encoding.html
