+++ assessments/tests/ressources/score_encoding_sample.json
