+++ assessments/urls.py
@@ -1,6 +1,6 @@
 ##############################################################################
 #
-# OSIS stands for Open Student Information System. It's an application
+#    OSIS stands for Open Student Information System. It's an application
 #    designed to manage the core business of higher education institutions,
 #    such as universities, faculties, institutes and professional schools.
 #    The core business involves the administration of students, teachers,
@@ -15,7 +15,7 @@
 #
 #    This program is distributed in the hope that it will be useful,
 #    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 #    GNU General Public License for more details.
 #
 #    A copy of this license - GNU General Public License - is available
@@ -23,8 +23,10 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-from dashboard.models import score_encoding as mdl_score_encoding
-import json
-import datetime
+from django.conf.urls import url
+from .views import score_encoding
 
-
+urlpatterns = [
+    url(r'^scores_encoding/$', score_encoding.score_encoding, name='scores_encoding'),
+    url(r'^scores_encoding/download/$', score_encoding.download_papersheet, name='scores_download')
+]
