+++ dashboard/models/__init__.py
@@ -1,1 +1,1 @@
-from dashboard.models import score_encoding
+
