+++ frontoffice/urls.py
@@ -63,6 +63,8 @@ if 'exam_enrollment' in settings.INSTALLED_APPS:
     urlpatterns = urlpatterns + (url(r'^exam_enrollment/', include('exam_enrollment.urls')), )
 if 'attestation' in settings.INSTALLED_APPS:
     urlpatterns = urlpatterns + (url(r'^attestation/', include('attestation.urls')), )
+if 'assessments' in settings.INSTALLED_APPS:
+        urlpatterns = urlpatterns + (url(r'^assessments/', include('assessments.urls')),)
 
 handler404 = 'base.views.common.page_not_found'
 handler403 = 'base.views.common.access_denied'
